Marjor Multi Refactor
Adds a new struct which keeps track of objects on the server.  When a client fires, the client sends the physics of the shot info to the server. The server then adds the weapon wherever it would have been if the client were in the right position. Woohoo!....

Just need a little more testing to make sure that the *one* bug I found is fixed.

Hope you like it.

Object Packet Major Refactor:

Change 1) Reduce Subsystem update overhead:

Allows us to only send subsystems that are required, not everything.
On the whole this should save a lot of bandwidth, allowing us to add a few much needed bits to other parts of the object update packet

I added a flag in all places that subsystems are changed in mission so that the engine would know which to send.

Change 2)  Foward Thrust:

Now actually use the forward thrust we send over the network.  Previously, it was sent, read, and then ignored.

Change 3) - Add Side Thrust:

Include side thrust!  Now those Mara's and Dragons should be where you expect. At the cost of only one more byte per packet. (if position is flagged, like it almost always is)

Change 4) - Change Order For Vel Calc:

Change positions of veloctiy and orientation in the packet so that correct orienation can take part in retreiving the correct velocity from compression within the packet.

Change 5) - Increase velocity accuracy
Decrease the compression of velocity in the packet. Uses only one more byte but should allow the velocity to be calculated to be within .5 m/s of the server velocity instead of 3 m/s.

Change 6) - Increase size of Sequence Number by One Byte

Add a byte to the object update sequence number (change it from a ubyte to a ushort). This should make object updating smoother since wrapping the 16 bit sequence number is much, much rarer