Make sure Rearm Limists Are Set or Reinstated When changin Weapons
Any function that changes the weapon should also change the rearm limit.  This should ensure that wl_bash_ship_weapons changes the rearm limit.

This could be a fix to the rearm bug, since this function is often called right after weapons are changed to the default for unrelated reasons, to undo those changes.

This portion of the code really needs a refactor though.  It took three days to figure out what was going on.