Increase accuracy of Velocity Transmission
Previously the velocity being transmitted  would be either within 1m/s or up to 3m/s off.  To figure this out, I recorded the inputs of the writing portion of the function and the outputs of the reading part of the function to find the discrepency.

This moves the velocity so that the orientation from the server can take part in its calculation.  Best way to do this was to split the and simplify the two functions so that we wouldn't need a myriad of function arguments.

Make this pr more understandable by splitting velocity bashing to a new function

Allow the communicated forward thrust to actually be part of the interpolation calculations.