Start Adding Infrastructure for lua packets
beginning to write the functions and manager that will take care of the majority of lua packets

save work....

just saving tiny progress

Begin change to more generic system

Adjust how HEADER_LENGTH is declared so that it can be used in a constexpr

We are now setting up shop in multimsgs to allow the code to be more modular.

Ok.  Some mediocre progress with a more verbose system